+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 ## 
+- 2025-05-24 21:14: Added Lodging and Restaurant section
 - 2025-05-24 19:55: Fixed render issue with scrolling image
 - 2025-05-24 19:50: Added background scroll animation to hero section
 - 2025-05-24 17:24: Build #2